refactor: split the effort tests and add's engine fields onto their own seams
api-handlers-more.test.ts crossed the file-size cap at 553 lines. The seam is
real: that file pins the generic simple-RPC edit verbs, where the payload IS
the behavior, while the set-effort cases are about engine RESOLUTION. They move
to api-set-effort.test.ts, beside api-add-effort.test.ts, which covers the same
gate on the create side.

handlers-add.ts was at 473 with a cap warning after this branch added to it, so
the engine half — which command, which protocol, which effort level — moves to
add-engine-fields.ts. handlers-add.ts keeps the create orchestration.
